seperate random impl from mathnc
