+++ ENIGMAsystem/SHELL/Universal_System/mathnc.h
@@ -42,7 +42,6 @@
 #endif
 
 #include "math_consts.h"
-#include "mathrand.h"
 #include "scalar.h"
 #include "var4.h"
